backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
openai==1.3.5
anthropic==0.7.1
google-generativeai==0.3.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
firebase-admin==6.2.0
requests==2.31.0
streamlit==1.28.1
